First changes to modify subset list mode organisation into entry order subsets. NOT WORKING
Previous to this, the method of selecting subsets was to rebin the data into geometric views and set the views. Here I try to reformat the subset selection with list mode data by first counting the number of events recorded. Divide the total by the number of subsets to determine the number of events per subset. Iterating through the records of subsets, determine whether the record number is within the subset range, if so, perform the projections; if not, and we are less than the subset min, go to next; if more than subset max, break out of update because there will be no more events in this subsets.

This method is slightly flawed because the greater the subset num, the more records the algorithm searches through.

NOT WORKING
Currently this implementation does not actually provide any update and the OSMAPOSL images result in being blank zeros. I will investigate what change I implemented that has caused this by comparing to the master branch.