Basic working version of the block subset sampling method
This method is limited in it implementation because it searches through all the data to find the subset number. After num_events_investiaged > subset_event_max then it breaks because no more events are present.

Fixed the previous issue where the image was becoming zero. A result of how I was "resetting" record list.